requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
pydantic
python-dotenv
google-generativeai
